requirements.txt
django
Pillow
colorama
prompt_toolkit
python-dotenv
python-decouple
requests
